Relicense this repository (#1702)
* Relicense this repository

This commit relicenses most crates in this repository under a new
license:

    Apache-2.0 WITH LLVM-exception OR Apache-2.0 OR MIT

where the previous license was:

    Apache-2.0 WITH LLVM-exception

This is done in accordance with #1637 with signoff from all known
stakeholders with copyrightable contributions. The tl;dr; for why this
change is being made is to better make these crates compatible with the
Rust ecosystem by providing `Apache-2.0 OR MIT` as an option.

Notably this does NOT relicense the C API due to the original
contributor not being able to be reached for the relicensing here. That
part of this repository still retains its original license.

Closes #1637

* Remove license symlinks