Add predefined `WasmFeatures` sets for wasm 1.0/2.0
This commit adds new `WasmFeatures::wasm{1,2}()` function which returns
the set of features enabled for those versions of the WebAssembly
specification. This is then reflected on the CLI with `--features wasm1`
or `--features wasm2` or `--features mvp`.

This is intended to make it a bit easier to run the validator on the CLI
for the MVP version of wasm without passing a long set of
`-disable-this-feature` options.

This also updates the help text for `wasm-tools validate` to be a bit
more readable.