threads: round out `missing-features` tests
This change attempts to add tests that check that
shared-everything-threads features indeed fail when the proposal is not
enabled. In doing this, I realized that a few features fail on other
features first (GC, reftypes), which meant the tests couldn't be as
exhaustive as I would have liked. This commit also changes the wording
of one of the failures to be consistent across the proposal.