Test `wasm_encoder::reencode` in roundtrip tests (#1720)
* Update wasm-encoder's type encoding to match `wast`

This updates it to not only match the structure but also binary output.
This means that shorter forms are used a bit more and encodings for
shared-everything-threads are additionally fixed.

* Update wasm_encoder's reencoding of const exprs

Use more generic helpers to support multiple instructions and
additionally more instructions. This refactors the `parse_instruction`
helper to return the instruction directly instead of unconditionally
encoding into a `Function`.

* Simplify some type encoding in `wast`

No functional change here, just reducing some duplication of code.

* Test wasm-encoder roundtrips of `roundtrip` tests

This commit updates the wasm-roundtrip test suite to additionally run
all binaries through the reencoding phase of `wasm-encoder` to assert
the final binary matches `wast`. This is skipped for components for now
as reencoding support isn't here for components just yet.

* Fix error message when wasm binaries differ

The condition here was inverted by accident meaning that more helpful
error messages were not being generated.