Finish reencoding support for components to wasm-encoder  (#1722)
* Extend re-encode functionality to support components

Signed-off-by: itowlson <ivan.towlson@fermyon.com>

* WIP: indexes, reorg

Signed-off-by: itowlson <ivan.towlson@fermyon.com>

* Finish reencoding support for components to wasm-encoder

This commit is an extension of #1656 to support components in the
reencoding support of wasm-encoder. The precise shape of the trait will
probably change over time but this is at least a first stab at getting
things working to map all the wasmparser types to wasm-encoder for components.

All existing support for components in `reencode.rs` was moved to a new
`ReencodeComponent` trait which inherits from the already-existing
`Reencode` trait. This is all located in a new `component.rs` file
beneath `reencode.rs` although crate-wise it's all showing up in the
same `wasm_encoder::reencode` module.

* Fix compat with Rust 1.78

---------

Signed-off-by: itowlson <ivan.towlson@fermyon.com>
Co-authored-by: itowlson <ivan.towlson@fermyon.com>