Handle `ref.eq` types from unreachable code
This validator's existing logic assumed that the bottom and heap bottom
types matched any reference type (see `_pop_operand`). This change
surfaces that logic to `visit_ref_eq` as well: if `pop_maybe_shared_ref`
returns `None` (a bottom or heap bottom) for either value to `ref.eq`,
we skip checking the shared-ness match, assuming like we do in
`_pop_operand` that a bottom or heap bottom matches any reference type
(we've already checked that the types are subtypes of `eqref`).