Add a feature for disabling GC types
This commit adds support for a new wasm feature named
`WasmFeatures::GC_TYPES`. This does not correspond to any upstream
proposal and is intended for use in Wasmtime for disabling the runtime
garbage collector at compile time. This serves as a finer-grained switch
to disable the runtime dependency at validation time on a garbage
collector without disabling all the features that were added in other
proposals. For example the `reference-types` proposal also added support
for multi-table which disabling a runtime garbage collector doesn't need
to disable.