Remove `Display for {Alignment,ArchitectureSize}`
These display representations are easy to get confused with the previous
return values of `u32` which do implement `Display`, but the `Display`
here isn't as applicable as it used to be. As a stopgap remove the
`Display` implementations entirely to help identify locations that need
to be updated to explicitly use an accessor or otherwise to calculate
the size/align information.