refactor: Remove optional feature attribute from stable feature gates (#1741)
* Remove feature flag from post-stabilization feature gates

This commit removes the optional `feature` specification from feature
gates (`@since`, in particular).

This change should simplify the usage of feature gates (see
https://github.com/WebAssembly/component-model/pull/387) for more
discussion.

This is a breaking change for those who were depending on `wit-parser`
as `feature` now no longer present.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

* Update wit-component to not use feature option in since

This commit updates `wit-component` to remove reliance on the
`feature` option when dealing with post-stabilization (`@since`)
feature gates.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

* Update tests to remove optional feature on since gates

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

---------

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>