Replace `type_of_function` with `type_index_of_function`
`WasmModuleResources` previously retrieved the entire `FuncType` but
this obscured the sharedness of the function in question, since the
shared flag is held a level up in `CompositeType`. By replacing
`WasmModuleResources::type_of_function` with a more low-level
`WasmModuleResources::type_index_of_function`, we can reuse the existing
helper functions that retrieve a `FuncType` from a `CompositeType` and
do the necessary sharedness checks.