threads: no shared function expansion (#1745)
If I understand `wast`'s `Expander` correctly (which I may not!), there
is no need for it to know how to create type definitions for shared
functions because there is no way to write WAT that would mark a
function shared in an inline way. Since the only way to create a shared
function is with an explicit type ID (e.g., `(func (type $ft))`), we can
set `shared = false` when a `TypeDef` must be created.