Don't do "sloppy" error matching on our own tests (#1755)
Right now all upstream spec tests are asserted to have roughly the same
shape of error as wasmparser produces to ensure that wasmparser isn't
failing for the wrong reason on an invalid wasm module. This involves
quite a large function, however, which is quite fuzzy on how it matches
error strings. None of this logic should be applicable in the
`tests/local/**/*.wast` tests however because those are all checked in
to this repository and aren't intended to reflect upstream error
messages.

This commit disables the sloppy error matching and requires that the
exact error contains the exact string in the assertion. This then goes
through some tests and updates some of the expected errors.