Generate WITs with the latest `wit-smith` (#1758)
With the new wit64 fuzzer don't use the old `wit-smith` but instead use
the newer version. The old version still generates multi-return
functions which are no longer supported by default in WIT. This will
eventually cause a problem when the newest wit-smith generates features
the "old" version no longer supports, but I figure that's a bridge that
can be crossed when we get there.