Add a shorthand feature constructor for Wasm 3.0
This is still in-development but is likely close to the final set of
proposals. Additionally move the previous function constructors to
associated `const` values to be a bit more idiomatic with other feature
names. Finally refactor the CLI to more easily support these named sets
in suggestions and refactor some feature selection in the roundtrip tests.