Fix running tests with `RUST_BACKTRACE=1`
The previous version of test output used the `:?` formatter for
`anyhow::Error` which would include the native backtrace when
`RUST_BACKTRACE=1` is set. This commit updates formatters to use `:#`
which is a bit less readable but doesn't include the backtrace when this
env var is set. Additionally a test is added to CI to ensure this
doesn't regress in the future.

This is adapted from #1578.

Closes #1578