Proceed to the next step in the "remove `interface`" transition (#1753)
* Proceed to the next step in the "remove `interface`" transition

Starting in WebAssembly/component-model#263 the component model binary
specification was updated in a technically breaking way to encode
binaries differently. This was intended to be rolled out in a manner
that minimized breakage however so bytecodealliance/wasm-tools#1262
implemented validation where the prefix byte which changed was actually
ignored if it was 0 or 1. The encoder at the time still emitted 1 as the
prefix byte, however. The intention was that once the validator had
percolated enough the encoder would switch to using 0.

Unfortunately this change was basically forgotten about until now with
WebAssembly/component-model#391, but now's probably "enough time passed"
so the encoder is updated to emit a 0x00 leading byte in this situation
instead of the historical 0x01 byte. This will start the transition
period to eventually removing the validator's acceptance of the 0x01
byte.

* Update test expectation