Share more core type infrastructure in components (#1765)
This commit chiefly fixes the validator panic found in #1763. This is
done by sharing more infrastructure when validating types between the
core module validator and the component validator. This sharing is done
by extracting a new trait and moving methods from the core module to the
trait and then implementing the trait for both component and core
contexts.

When writing tests this additionally discovered that name resolution on
core types wasn't happening within components. The fix there was the
same as the core module validator which was to extract shared bits to a
trait and then implement the trait in both locations.

Closes #1763