Allow parsing `rectypes` in components (#1764)
* Allow parsing `rectypes` in components

This change supports [#392], which adds a way to use GC's `rectypes` as
type definitions in components. Previously, only function types were
supported and there was no way express array and struct types. This
keeps the previous function decoding support based on peeking the
function type `0x60` prefix but adds support for encoding `rectypes`
with a new `0x00` prefix.

[#392]: https://github.com/WebAssembly/component-model/pull/392

Co-authored-by: Alex Crichton <alex@alexcrichton.com>

* Apply `0x00` prefix to non-final `sub`; add tests

This follows along with the most recent discussion in the component
model PR ([#392]).

[#392]: https://github.com/WebAssembly/component-model/pull/392

Co-authored-by: Alex Crichton <alex@alexcrichton.com>

* review: keep variant as `ComponentCoreTypeId::Sub`

* review: remove leftover comment

* review: remove resolved TODOs

* review: move `From` implementations to `core/binary.rs`

* review: remove `parse_component_sub_type`

---------

Co-authored-by: Alex Crichton <alex@alexcrichton.com>