Remove the ability to turn off semicolons in WIT
This is the final step in the transition to requiring semicolons in WIT.
Long ago semicolons weren't required in WIT and now they are. This
completes the last step of the transition after #1269 by removing the
`WIT_REQUIRE_SEMICOLONS` env var and making it a hardcoded requirement
that they're present.