Start requiring f32/f64 in WIT by default
This commit continues the transition started in #1364 to rename the
`float32 `and `float64` types in WIT to `f32` and `f64`. Now that
support has been present for some time to parse `f32` and `f64` this
commit flips the WIT parser to requiring it by default. The error
message for `float32` and `float64` has additionally been updated to
explain a bit more about this transition.