Beef up validity assertions on `Resolve`
This commit extends the `Resolve::assert_valid` function with more
assertions about the structure of worlds notably to guarantee they are
"elaborated" meaning that they always list all dependencies of imports.
This is required by bindings generators and encoding. This property is
already upheld internally and is intended to reflect a preexisting
property with dynamic assertion checks.

The underlying motivation for this is to assist in the development and
fuzzing of #1784.