json-from-wast: Include binary form of text files where possible (#1775)
This commit is aimed at addressing #1771 by including a new
`binary_filename` field in the JSON which is populated when the test
itself is `module quote`, or a textual module, but the module parse
successfully. This should help avoid the need for a text parser for
external users using the tool while still reflecting the test case
itself where it's text-based.

Closes #1771