Deduplicate how many times type-related opcodes are listed
This commit refactors the parsing of core types in wasmparser to
deduplicate the number of times that opcodes for particular types are
listed. Previously each layer of parsing types would check for all valid
prefixes and then delegate to a "lower" parser if applicable. This meant
though that a type's opcode was listed at each layer it could be parsed,
resulting in duplication.

Instead now types only parse what is exclusive to that layer and then a
lower layer parses everything else. The only gotcha with this is to
preserve reasonable error messages (lest every error become "invalid
abstract heap type"). For this a new `BinaryReaderErrorKind` enum
(internal) was added which is used to track this and the message is
updated as parsing fails and errors progress upwards.