Fix an infinite loop in wasm-smith
This fixes a possible infinite loop in wasm-smith when the gc proposal
is enabled. When the GC proposal is enabled and a minimum number of
types were specified it could get wasm-smith stuck in an infinite loop
where it would either always create an empty rec group or it would
always try to clone an empty rec group. The fix here is to thread a
boolean which toggles whether an empty rec group is allowed based on
whether we're generating the minimum number of types or any extra types.

This is a relatively old bug so I'm not sure why this hasn't been
discovered prior to this. Local fuzzing in Wasmtime found this quite
quickly, so I may have just gotten unlucky. We also tend to not look at
timeouts on OSS-Fuzz that closely as well, so perhaps this is a reminder
to actually do that.