Enhance `print` and `validate` fuzzers
This commit updates these two fuzzers, which just have random bytes
thrown at them, to additionally fuzz various config options in
`wasmprinter`. The `validate` fuzzer in particular is going to be
interesting since it creates "halfway invalid" modules where the
structure is valid and random bytes may be thrown in a function body.
That should stress various bits and pieces of the internals of
`wasmprinter` in addition to `wasmparser`.